requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0
matplotlib>=3.6.0
seaborn>=0.12.0
joblib>=1.2.0
pyyaml>=6.0
requests>=2.28.0
notebook>=7.0.0
ipykernel>=6.0.0
scipy>=1.10.0
